- created DBConnection class to connect to database. Created ApplicationDAO to interact with database.
- created new folder 'libs' to paste the MySQL connector into.
- Then added the connector to the build path. got added under Reference
Libraries.

- added a UUID UserID to the User class, also added setter methods for
all variables and a getter & setter method for the password.

- added a second User constructor to populate User and then use the set
methods to add content from the database <<< can be replaced with a
creational pattern in the next iteration of the project.

- added a 'roleName' String under user to get and pass the name of the
role to the database. The role name will have to be used to set the
actual role as the software develops.